+++ packages/react-start-client/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @tanstack/react-start-client
 
+## 1.166.42
+
+### Patch Changes
+
+- Updated dependencies [[`8b97002`](https://github.com/TanStack/router/commit/8b97002af3f6d15204e60c55d3f5735b78bd7efe)]:
+  - @tanstack/start-client-core@1.167.19
+
+## 1.166.41
+
+### Patch Changes
+
+- Updated dependencies [[`4d864ee`](https://github.com/TanStack/router/commit/4d864eebbd184265eabb563d326ab409c93feb17)]:
+  - @tanstack/react-router@1.168.24
+  - @tanstack/router-core@1.168.16
+  - @tanstack/start-client-core@1.167.18
+
 ## 1.166.40
 
 ### Patch Changes