+++ packages/react-start/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @tanstack/react-start
 
+## 1.167.48
+
+### Patch Changes
+
+- Updated dependencies [[`8b97002`](https://github.com/TanStack/router/commit/8b97002af3f6d15204e60c55d3f5735b78bd7efe)]:
+  - @tanstack/start-client-core@1.167.19
+  - @tanstack/react-start-client@1.166.42
+  - @tanstack/react-start-rsc@0.0.27
+  - @tanstack/react-start-server@1.166.43
+  - @tanstack/start-plugin-core@1.169.4
+  - @tanstack/start-server-core@1.167.21
+
+## 1.167.47
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-plugin-core@1.169.3
+  - @tanstack/react-start-rsc@0.0.26
+
+## 1.167.46
+
+### Patch Changes
+
+- Updated dependencies [[`4d864ee`](https://github.com/TanStack/router/commit/4d864eebbd184265eabb563d326ab409c93feb17)]:
+  - @tanstack/react-router@1.168.24
+  - @tanstack/start-plugin-core@1.169.2
+  - @tanstack/start-server-core@1.167.20
+  - @tanstack/react-start-client@1.166.41
+  - @tanstack/react-start-rsc@0.0.25
+  - @tanstack/react-start-server@1.166.42
+  - @tanstack/start-client-core@1.167.18
+
+## 1.167.45
+
+### Patch Changes
+
+- Updated dependencies [[`9252206`](https://github.com/TanStack/router/commit/9252206e5aeafe53e31eb7baa491d07a597c4dc6)]:
+  - @tanstack/start-plugin-core@1.169.1
+  - @tanstack/react-start-rsc@0.0.24
+
 ## 1.167.44
 
 ### Patch Changes