+++ packages/router-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tanstack/router-generator
 
+## 1.166.34
+
+### Patch Changes
+
+- Updated dependencies [[`4d864ee`](https://github.com/TanStack/router/commit/4d864eebbd184265eabb563d326ab409c93feb17)]:
+  - @tanstack/router-core@1.168.16
+
 ## 1.166.33
 
 ### Patch Changes