+++ packages/solid-router/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tanstack/solid-router
 
+## 1.168.21
+
+### Patch Changes
+
+- Add TanStack Start inline CSS manifest support for SSR so route styles can be embedded in the HTML response and hydrated without duplicate stylesheet links. ([#7253](https://github.com/TanStack/router/pull/7253))
+
+- Updated dependencies [[`4d864ee`](https://github.com/TanStack/router/commit/4d864eebbd184265eabb563d326ab409c93feb17)]:
+  - @tanstack/router-core@1.168.16
+
 ## 1.168.20
 
 ### Patch Changes