+++ packages/start-storage-context/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tanstack/start-storage-context
 
+## 1.166.30
+
+### Patch Changes
+
+- Updated dependencies [[`4d864ee`](https://github.com/TanStack/router/commit/4d864eebbd184265eabb563d326ab409c93feb17)]:
+  - @tanstack/router-core@1.168.16
+
 ## 1.166.29
 
 ### Patch Changes